[#1375] Revise public API
access cycle implementation is not part of api, goes to internal package
