+++ bundles/org.eclipse.passage.lic.base/META-INF/MANIFEST.MF
@@ -9,7 +9,7 @@ Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Require-Bundle: org.eclipse.passage.lic.api;bundle-version="0.0.0";visibility:=reexport
 Export-Package: org.eclipse.passage.lic.base,
- org.eclipse.passage.lic.base.access,
+ org.eclipse.passage.lic.base.access;x-internal:=true,
  org.eclipse.passage.lic.base.acquire,
  org.eclipse.passage.lic.base.agreements;x-friends:="org.eclipse.passage.lic.execute",
  org.eclipse.passage.lic.base.conditions,
@@ -25,6 +25,7 @@ Export-Package: org.eclipse.passage.lic.base,
  org.eclipse.passage.lic.base.version,
  org.eclipse.passage.lic.internal.base;x-friends:="org.eclipse.passage.lic.equinox.tests",
  org.eclipse.passage.lic.internal.base.access;x-internal:=true,
+ org.eclipse.passage.lic.internal.base.access.storage;x-internal:=true,
  org.eclipse.passage.lic.internal.base.conditions;x-internal:=true,
  org.eclipse.passage.lic.internal.base.inspection.hardware;x-friends:="org.eclipse.passage.loc.licenses.core,org.eclipse.passage.lic.oshi",
  org.eclipse.passage.lic.internal.base.requirements;x-internal:=true,
