+++ tests/org.eclipse.passage.lic.base.tests/src/org/eclipse/passage/lic/internal/base/tests/requirements/RequirementsTest.java
@@ -23,10 +23,10 @@ import org.eclipse.passage.lic.api.registry.StringServiceId;
 import org.eclipse.passage.lic.api.requirements.Requirement;
 import org.eclipse.passage.lic.api.requirements.ResolvedRequirements;
 import org.eclipse.passage.lic.api.tests.fakes.requirements.FakeResolvedRequirements;
-import org.eclipse.passage.lic.base.access.Requirements;
 import org.eclipse.passage.lic.base.diagnostic.code.NoRequirements;
 import org.eclipse.passage.lic.base.diagnostic.code.NoServicesOfType;
 import org.eclipse.passage.lic.base.registry.ReadOnlyRegistry;
+import org.eclipse.passage.lic.internal.base.access.Requirements;
 import org.junit.Test;
 
 @SuppressWarnings("restriction")
